Reuse processor test settler for shutdown start tests
